backend/requirements.txt
# Shared Backend Dependencies
# This file contains common dependencies used across all backend services

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
cryptography==41.0.8

# AWS & DynamoDB
boto3==1.34.0
botocore==1.34.0

# HTTP Client
requests==2.31.0

# Configuration Management
python-dotenv==1.0.0

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
